+++ lib/solargraph/shell.rb
@@ -74,7 +74,11 @@ module Solargraph
     desc 'download-core [VERSION]', 'Download core documentation'
     def download_core version = nil
       ver = version || Solargraph::YardMap::CoreDocs.best_download
-      puts "Downloading docs for #{ver}..."
+      if RUBY_VERSION != ver
+        puts "Documentation for #{RUBY_VERSION} is not available. Reverting to closest match..."
+      else
+        puts "Downloading docs for #{ver}..."
+      end
       Solargraph::YardMap::CoreDocs.download ver
       # Clear cached documentation if it exists
       FileUtils.rm_rf Dir.glob(File.join(Solargraph::YardMap::CoreDocs.cache_dir, ver, '*.ser'))